# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a Web Push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionBody"
      responses:
        "200":
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/unsubscribe:
    delete:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a Web Push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeBody"
      responses:
        "200":
          description: Subscription removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/subscription-status:
    get:
      operationId: getPushSubscriptionStatus
      tags: [push]
      summary: Check whether the current user has any active push subscription
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionStatus"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /push/prefs:
    get:
      operationId: getPushPrefs
      tags: [push]
      summary: Get current user's push notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushPrefs"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePushPrefs
      tags: [push]
      summary: Update push notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushPrefsUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushPrefs"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/test:
    post:
      operationId: sendPushTest
      tags: [push]
      summary: Send a sample push notification to the calling user's subscribed devices
      description: |
        Lets a signed-in user verify their phone actually pops up an OS-level
        notification. Sends to every subscription endpoint registered for the
        caller. Per-user rate limited so a misbehaving client cannot spam
        their own devices.
      responses:
        "200":
          description: Test push dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User has no push subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token returned by login or registration.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: Browser session cookie.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PushSubscriptionKeys:
      type: object
      properties:
        p256dh:
          type: string
        auth:
          type: string
      required:
        - p256dh
        - auth
    PushSubscriptionBody:
      type: object
      properties:
        endpoint:
          type: string
          format: uri
        keys:
          $ref: "#/components/schemas/PushSubscriptionKeys"
      required:
        - endpoint
        - keys
    PushUnsubscribeBody:
      type: object
      properties:
        endpoint:
          type: string
          format: uri
      required:
        - endpoint
...
    PushSubscriptionStatus:
      type: object
      properties:
        subscribed:
          type: boolean
      required:
        - subscribed
    PushPrefs:
      type: object
      properties:
        pushExpiry:
          type: boolean
        pushBroadcast:
          type: boolean
        pushDailySummary:
          type: boolean
        pushMarketNews:
          type: boolean
        pushCalendarEvents:
          type: boolean
        pushPriceAnomaly:
          type: boolean
        pushWeeklyRecap:
          type: boolean
        pushSignalFlip:
          type: boolean
        marketOpenSessions:
          type: array
          items:
            type: string
            enum: [tokyo, london, newyork]
          description: FX sessions the user wants a 5-min pre-open ping for. Empty = off.
        pushDormancyNudge:
          type: boolean
          description: Opt-in toggle for the weekly "we miss you" nudge after 7+ days idle.
        pushOnboarding:
          type: boolean
          description: One-shot 24h-after-signup empty-watchlist nudge.
        disengageNoticeCategory:
          type: string
          nullable: true
          description: When non-null, the UI should render a one-time banner explaining auto-pause.
        guardrailRevenge:
          type: boolean
          description: Show soft warning when a loss on this instrument fired within the revenge window.
        guardrailOvertrading:
          type: boolean
          description: Show soft warning when the user crosses the per-hour or per-day analysis count.
        guardrailHighRisk:
          type: boolean
          description: Show soft warning when a high-impact event for the instrument prints within 30 min.
        coolingOffEnabled:
          type: boolean
          description: Opt-in 30-minute countdown after a significant loss before showing the analyse button warning.
        pushAnalysisCompleted:
          type: boolean
          description: OS push when an AI analysis finishes processing.
        pushTpSlHit:
          type: boolean
          description: OS push when a tracked trade plan's TP or SL level is crossed.
        pushLoginAlert:
          type: boolean
          description: OS push on a new login. The in-app notification is always created regardless of this toggle.
        nativePushEnabled:
          type: boolean
          description: Master switch for native (FCM) push delivery to registered mobile devices.
        webPushEnabled:
          type: boolean
          description: Master switch for Web Push (VAPID) delivery to subscribed browsers.
        quietHoursEnabled:
          type: boolean
          description: When false, quiet hours are disabled entirely for this user.
        quietHoursStart:
          type: string
          description: HH:MM 24h local time quiet hours begin.
          example: "22:00"
        quietHoursEnd:
          type: string
          description: HH:MM 24h local time quiet hours end.
          example: "07:00"
        notificationTimezone:
          type: string
          description: IANA timezone quietHoursStart/quietHoursEnd are interpreted in.
          example: Asia/Jakarta
        progressionNotificationsEnabled:
          type: boolean
          description: Enable personal progression notifications.
      required:
        - pushExpiry
        - pushBroadcast
        - pushDailySummary
        - pushMarketNews
        - pushCalendarEvents
        - pushPriceAnomaly
        - pushWeeklyRecap
        - pushSignalFlip
        - marketOpenSessions
        - pushDormancyNudge
        - pushOnboarding
        - guardrailRevenge
        - guardrailOvertrading
        - guardrailHighRisk
        - coolingOffEnabled
        - pushAnalysisCompleted
        - pushTpSlHit
        - pushLoginAlert
        - nativePushEnabled
        - webPushEnabled
        - quietHoursEnabled
        - quietHoursStart
        - quietHoursEnd
        - notificationTimezone
        - progressionNotificationsEnabled
    PushTestResult:
      type: object
      properties:
        delivered:
          type: integer
          description: Number of subscription endpoints the test push was dispatched to
      required:
        - delivered
...
    PushPrefsUpdate:
      type: object
      properties:
        pushExpiry:
          type: boolean
        pushBroadcast:
          type: boolean
        pushDailySummary:
          type: boolean
        pushMarketNews:
          type: boolean
        pushCalendarEvents:
          type: boolean
        pushPriceAnomaly:
          type: boolean
        pushWeeklyRecap:
          type: boolean
        pushSignalFlip:
          type: boolean
        marketOpenSessions:
          type: array
          items:
            type: string
            enum: [tokyo, london, newyork]
        pushDormancyNudge:
          type: boolean
        pushOnboarding:
          type: boolean
        dismissDisengageNotice:
          type: boolean
          description: Pass true to clear the one-time auto-pause banner.
        guardrailRevenge:
          type: boolean
        guardrailOvertrading:
          type: boolean
        guardrailHighRisk:
          type: boolean
        coolingOffEnabled:
          type: boolean
        pushAnalysisCompleted:
          type: boolean
        pushTpSlHit:
          type: boolean
        pushLoginAlert:
          type: boolean
        nativePushEnabled:
          type: boolean
        webPushEnabled:
          type: boolean
        quietHoursEnabled:
          type: boolean
        quietHoursStart:
          type: string
          description: HH:MM 24h local time.
          example: "22:00"
        quietHoursEnd:
          type: string
          description: HH:MM 24h local time.
          example: "07:00"
        notificationTimezone:
          type: string
          description: IANA timezone.
          example: Asia/Jakarta
        progressionNotificationsEnabled:
          type: boolean
...